feat: saveResult
